[coq] Fix warnings and try to handle primitive integers
We fix a couple of OCaml warnings, and in particular we try to handle
the new kernel constructor introduced in 8.10

I am not sure this is the right way to do it, please review.

Before merging this PR, a 8.9 branch should be created as this change
is not backwards compatible.